Combine light_api_util and lights_util into lights_util
